pyproject.toml
[project]
name = "cc-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.2",
    "greenlet>=3.3.2",
    "groq>=1.1.2",
    "livekit-agents[deepgram,openai,silero,turn-detector]~=1.4",
    "livekit-plugins-groq>=1.5.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "sse-starlette>=3.3.3",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"src/agent/prompts.py" = ["E501"]

[project.scripts]
lint = "ruff check --fix src/ tests/"
fmt = "ruff format src/ tests/"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
